changes to gps filter on state machine

+++ Control.ino
@@ -20,13 +20,13 @@ bool CheckEstimate()
 }
 
 // determine the altitude from Pressure and Temperature using Hypsometric formula
-long Pressure_Alt_Calc(long Pressure, long Temperature)
+float Pressure_Alt_Calc(float Pressure, float Temperature)
 {
-  long R = 287; // gas constant (J/kg*K)
-  long P0 = 101325; // atmospheric pressure at sea level (Pa)
+  float R = 287; // gas constant (J/kg*K)
+  float P0 = 101325; // atmospheric pressure at sea level (Pa)
 
-  long Pressure_Alt_SI = (R*Temperature/9.81)*log(P0/pressure); // returns altitude based on P and T in meters.
-  long Pressure_Alt = Pressure_Alt_SI*3.28084; // converts m to ft
+  float Pressure_Alt_SI = (R*Temperature/9.81)*log(P0/pressure); // returns altitude based on P and T in meters.
+  float Pressure_Alt = Pressure_Alt_SI*3.28084; // converts m to ft
 
   return Pressure_Alt;
 }
